# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API untuk sistem service motor. Sebagian besar endpoint membutuhkan token JWT (Bearer) yang diperoleh dari login/register.",
        "title": "Sistem Service Motor API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/service-records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil daftar record servis milik user yang login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Record"
                ],
                "summary": "Ambil semua record servis user",
                "responses": {
                    "200": {
                        "description": "Data record servis berhasil diambil",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat data record servis baru untuk user yang login",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Record"
                ],
                "summary": "Buat record servis",
                "parameters": [
                    {
                        "description": "Data record servis",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/example_com_my-api_internal_dto.CreateServiceRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Data record servis berhasil dibuat",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "request body tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "kendaraan atau jenis servis tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/service-records/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Mengambil data record servis milik user yang login berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Record"
                ],
                "summary": "Ambil record servis berdasarkan ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID record servis",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data record servis berhasil diambil",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "id tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "record servis tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data record servis milik user yang login berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Record"
                ],
                "summary": "Update record servis",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID record servis",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data record servis",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/example_com_my-api_internal_dto.UpdateServiceRecordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Data record servis berhasil diupdate",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "request body tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "record servis tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus data record servis milik user yang login berdasarkan ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Service Record"
                ],
                "summary": "Hapus record servis",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID record servis",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "berhasil hapus data record servis",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "id tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "token tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "record servis tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "example_com_my-api_internal_dto.CreateServiceRecordRequest": {
            "type": "object",
            "required": [
                "labor_cost",
                "notes",
                "odometer",
                "parts_cost",
                "service_date",
                "service_type_id",
                "vehicle_id",
                "workshop_name"
            ],
            "properties": {
                "labor_cost": {
                    "type": "number",
                    "minimum": 0
                },
                "notes": {
                    "type": "string"
                },
                "odometer": {
                    "type": "number",
                    "minimum": 0
                },
                "parts_cost": {
                    "type": "number",
                    "minimum": 0
                },
                "service_date": {
                    "type": "string"
                },
                "service_type_id": {
                    "type": "integer"
                },
                "vehicle_id": {
                    "type": "integer"
                },
                "workshop_name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
...
        "example_com_my-api_internal_dto.UpdateServiceRecordRequest": {
            "type": "object",
            "required": [
                "labor_cost",
                "notes",
                "odometer",
                "parts_cost",
                "service_date",
                "service_type_id",
                "vehicle_id",
                "workshop_name"
            ],
            "properties": {
                "labor_cost": {
                    "type": "number",
                    "minimum": 0
                },
                "notes": {
                    "type": "string"
                },
                "odometer": {
                    "type": "number",
                    "minimum": 0
                },
                "parts_cost": {
                    "type": "number",
                    "minimum": 0
                },
                "service_date": {
                    "type": "string"
                },
                "service_type_id": {
                    "type": "integer"
                },
                "vehicle_id": {
                    "type": "integer"
                },
                "workshop_name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token JWT dengan format: `Bearer \u003ctoken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}